# data status: optimize printing status list